requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.6
django-adminlte-3==0.1.6
django-bootstrap4==3.0.1
djangorestframework==3.12.4
mysqlclient==2.0.3
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
